+++ Libplanet.Explorer/Queries/ExplorerQuery.cs
@@ -28,8 +28,8 @@ namespace Libplanet.Explorer.Queries
                 "nodeState",
                 resolve: context => chainContext
             );
-            Field<NonNullGraphType<PolicyType<T>>>(
-                "policy",
+            Field<NonNullGraphType<BlockPolicyType<T>>>(
+                "blockPolicy",
                 resolve: context => chainContext.BlockChain.Policy
             );